moving all the android classes to the android package
