+++ src/main/java/com/novoda/notils/android/StrictModeManager.java
@@ -1,4 +1,4 @@
-package com.novoda.notils.logger.analyse;
+package com.novoda.notils.android;
 
 import android.os.StrictMode;
