+++ src/main/java/com/novoda/notils/android/toast/ToastDisplayers.java
@@ -1,4 +1,4 @@
-package com.novoda.notils.logger.toast;
+package com.novoda.notils.android.toast;
 
 import android.content.Context;
